Move reports under /nytprof and make work properly
Move the reporting under `/nytprof` - it wouldn't be surprising for an
app to have its own existing URLs under `/profile`.

Fixes to HTML-serving to serve all the files from the NYTProf HTML
output properly so you can drill down to subroutine reports etc, and so
the stylesheet etc is served.