Styling profiles list, use UTF8 for nytprof output
Style the profiles list a bit more nicely, and move the datetime column
after the more important ones - chances are you're probably going to be
skimming for the most recent execution of the route you're interested
in.  (Must add filtering...)

Use slurp_utf8 to read the nytprof HTML output so we output it correctly
without e.g. the microseconds symbol µ (mu) being ruined.